chore: cache chocolatey installs as recommended by Appveyor
We tried to save the Chocolatey cache in a previous PR, but it looks
like its not working properly. This PR implements the "correct" way to
cache Chocolatey installs, as recommended in Appveyor forums.

The arrows mean that the cache is invalidated if the file in the right
side (`appveyor.yml`) in this case, is changed.

See: http://help.appveyor.com/discussions/suggestions/592-speeding-up-chocolatey-downloads-and-installs
See: https://github.com/resin-io/etcher/commit/e603cb0838b005f1c8430bbce4c98b431d9c1ba9
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>