minifix(GUI): show Windows warning for images starting with W(8|9|10)
Some Windows images start with `W8`, `W9`, or `W10`. This commit extends
the `.looksLikeWindowsImage()` regular expression to catch these cases.

Since a `W` followed by a number can occur in other parts of a file
name, we only attempt to detect this pattern at the beginning of the
file name.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>