fix(GUI): handle spaces in installation path when elevating on Windows
Etcher will stuck in "Starting..." when executing the application on a
directory that contains spaces, like "C:\Program Files (x86)".

The problem is that the command is not quoted correctly when passed to
`cmd.exe /c`. This commit addresses the following specific problems:

- Quote the whole argument to `cmd.exe /c`
- Quote each individual argument after `call`

Change-Type: patch
Changelog-Entry: Fix application stuck at "Starting..." on Windows.
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>