chore: list all needed packages on the Dockerfile
At the moment, we only install the packages that we need but that are
missing from the base Docker image that we use.

For safety purposes (and to make it easier to switch to another base
image), we list all the needed packages, even if they usually come
pre-installed, since that would only be a no-op in the worst case.

See: https://github.com/resin-io/etcher/commit/296a55463714194d15c28eb79bb2939908f8cffd#commitcomment-22358869
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>