feat(sdk): implement usbboot mode
This commit installs `node-usb` v1.3.0 from GitHub, since that version
was never published to NPM, and is the only one that works with Visual
Studio 2015 (see https://github.com/tessel/node-usb/issues/109).

The usbboot communicates with a Raspberry Pi through USB and eventually
mounts it as a block device we can write to.

Change-Type: minor
Fixes: https://github.com/resin-io/etcher/issues/1541
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>